# Another fix for mills() overflow issue